DB on docker-compose

+++ db/schema.rb
@@ -2,8 +2,8 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# This file is the source Rails uses to define your schema when running `rails
-# db:schema:load`. When creating a new database, `rails db:schema:load` tends to
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
 # be faster and is potentially less error prone than running all of your
 # migrations from scratch. Old migrations may fail to apply correctly if those
 # migrations use external dependencies or application code.
@@ -296,7 +296,7 @@ ActiveRecord::Schema.define(version: 2021_07_20_121933) do
   end
 
   create_table "tito_discount_codes", force: :cascade do |t|
-    t.bigint "sponsorship_id", null: false
+    t.integer "sponsorship_id", null: false
     t.integer "kind", null: false
     t.string "tito_discount_code_id", null: false
     t.string "code", null: false
